logger class introduced
