+++ python/benchmarks/common.py
@@ -16,7 +16,8 @@ import os
 import bz2
 import urllib.request
 
-from arcticdb.util.environment_setup import TestLibraryManager, LibraryPopulationPolicy, get_console_logger
+from arcticdb.util.environment_setup import TestLibraryManager, LibraryPopulationPolicy
+from arcticdb.util.logger import get_logger
 
 ## You can now define SLOW ASV tests
 ## To run those tests also you have to set following OS variable
@@ -238,7 +239,7 @@ class AsvBase(ABC):
     """
     
     def get_logger(self) -> Logger:
-        return get_console_logger(self)    
+        return get_logger(self)    
     
     @abstractmethod
     def get_library_manager(self) -> TestLibraryManager:
