+++ python/tests/compat/arcticdb/test_lib_naming.py
@@ -1,6 +1,6 @@
 import pytest
 import sys
-from arcticdb.util.utils import get_logger
+from arcticdb.util.logger import get_logger
 from arcticdb_ext.exceptions import UserInputException
 from arcticdb.util.test import sample_dataframe
 from tests.util.mark import SLOW_TESTS_MARK, xfail_azure_chars
