+++ python/tests/conftest.py
@@ -8,7 +8,7 @@ As of the Change Date specified in that file, in accordance with the Business So
 
 import enum
 from typing import Callable, Generator, Union
-from arcticdb.util.environment_setup import get_console_logger
+from arcticdb.util.logger import get_logger
 from arcticdb.version_store._store import NativeVersionStore
 from arcticdb.version_store.library import Library
 import hypothesis
@@ -691,7 +691,7 @@ def _store_factory(lib_name, bucket, delete_bucket = True) -> Generator[Callable
         try:
             bucket.slow_cleanup()
         except Exception as e:
-            get_console_logger().warning(f"Exception caught during NativeVersionStore clear: {repr(e)}")
+            get_logger().warning(f"Exception caught during NativeVersionStore clear: {repr(e)}")
 
 
 @pytest.fixture
