+++ python/tests/enduser/test_authentication.py
@@ -6,7 +6,8 @@ import pytest
 import logging
 import re
 
-from arcticdb.util.utils import GitHubSanitizingException, get_logger
+from arcticdb.util.utils import GitHubSanitizingException
+from arcticdb.util.logger import get_logger
 from tests.util.mark import REAL_GCP_TESTS_MARK, REAL_S3_TESTS_MARK
 from tests.util.storage_test import real_gcp_credentials, real_s3_credentials
